move dev commands in own folder
